首页快速催收

+++ service/report/src/main/java/cn/fintecher/pangolin/report/model/PreviewTotalFollowModel.java
@@ -29,8 +29,8 @@ public class PreviewTotalFollowModel {
     @ApiModelProperty(notes = "本月催记数")
     private Integer currentMonthCount;
     @ApiModelProperty(notes = "在线时长")
-    private Double onlineTime;
+    private Integer onlineTime;
     @ApiModelProperty(notes = "离线时长")
-    private Double offlineTime;
+    private Integer offlineTime;
 
 }
